Chore: Move InvarFitting (#3468)
